✨(frontend) create hook useDate
Create hook useDate, it is an date helper.
We use it for now to format date.
