+++ src/frontend/apps/impress/src/hook/index.ts
@@ -1,0 +1,1 @@
+export * from './useDate';
